rewrite generate_index_file.py to make space for json index code
